Refactor NodeAPi to be in the right package and avoid code duplications
