+++ packages/bodiless-components/package.json
@@ -24,7 +24,7 @@
   },
   "scripts": {
     "build": "run-p copy build:lib",
-    "build:lib": "tsc -b ./tsconfig.json ./tsconfig.cjs.json",
+    "build:lib": "tsc -p ./tsconfig.json",
     "build:watch": "npm run build:lib -- --watch",
     "build:api-doc": "typedoc --options ../../typedoc.js --out doc/api ./src",
     "clean": "rimraf \"lib/*\" && rimraf tsconfig.tsbuildinfo && rimraf \"doc/api\"",
