+++ packages/bodiless-git/package.json
@@ -20,7 +20,7 @@
   },
   "scripts": {
     "build": "run-p build:lib",
-    "build:lib": "tsc -p ./tsconfig.json",
+    "build:lib": "tsc -b ./tsconfig.json ./tsconfig.cjs.json",
     "build:watch": "npm run build:lib -- --watch",
     "build:api-doc": "typedoc --options ../../typedoc.js --out doc/api ./src",
     "clean": "rimraf \"lib/*\" && rimraf tsconfig.tsbuildinfo && rimraf \"doc/api\""
@@ -31,7 +31,10 @@
     "enzyme": "^3.9.0",
     "informed": "^4.44.0",
     "lodash": "^4.17.19",
-    "universal-cookie": "^4.0.3"
+    "universal-cookie": "^4.0.3",
+    "isomorphic-git": "^1.10.1",
+    "find-up": "^4.1.0",
+    "node-cache": "^5.1.2"
   },
   "peerDependencies": {
     "react": "^18.2.0"
