+++ packages/bodiless-page/package.json
@@ -26,7 +26,7 @@
   },
   "scripts": {
     "build": "run-p build:lib",
-    "build:lib": "tsc -p ./tsconfig.json",
+    "build:lib": "tsc -b ./tsconfig.json ./tsconfig.cjs.json",
     "build:watch": "npm run build:lib -- --watch",
     "build:api-doc": "typedoc --options ../../typedoc.js --out doc/api ./src",
     "clean": "rimraf \"lib/*\" && rimraf tsconfig.tsbuildinfo && rimraf \"doc/api\""
