+++ packages/bodiless-page/src/NodeApi/findComponentPath.ts
@@ -1,0 +1,32 @@
+/**
+ * Copyright © 2023 Johnson & Johnson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import path from 'path';
+import fs from 'fs';
+
+/**
+ * Helper function to find page component.
+ * @param  {...string} pathSegments Path to component directory.
+ */
+export const findComponentPath = (...pathSegments: string[]): string | null => {
+  let componentPath;
+  // Allowed component extensions are jsx, tsx and json.
+  ['index.jsx', 'index.tsx', 'index.json'].some(item => {
+    componentPath = path.resolve(...pathSegments, item);
+    if (fs.existsSync(componentPath)) {
+      return true;
+    }
+    return false;
+  });
+  return componentPath || null;
+};
