+++ packages/gatsby-theme-bodiless/gatsby-node.js
@@ -21,10 +21,18 @@
 const pathUtil = require('path');
 const fs = require('fs');
 const webpack = require('webpack');
-const { getDisabledPages } = require('@bodiless/components/lib/cjs/NodeApi');
+const {
+  getDisabledPages,
+  findComponentPath,
+  findSubPageTemplateTemplate,
+  findTemplate,
+} = require('@bodiless/page/lib/cjs/NodeApi');
+const {
+  createGitInfo,
+} = require('@bodiless/git/lib/cjs/NodeApi');
 const { createFilePath } = require('gatsby-source-filesystem');
 const { addStaticReplacementPlugin } = require('@bodiless/webpack');
-const { onCreateNode, createSlug, createGitInfo } = require('./create-node');
+const { onCreateNode, createSlug } = require('./create-node');
 const createRedirectAlias = require('./create-redirect-alias');
 const Logger = require('./Logger');
 // eslint-disable-next-line import/order
@@ -51,74 +59,6 @@ exports.onCreateBabelConfig = args => {
   setBabelPlugin({
     name: 'babel-plugin-preval',
   });
-};
-
-/**
- * Helper function to find page component.
- * @param  {...any} pathSegments Path to component directory.
- */
-const findComponentPath = (...pathSegments) => {
-  let componentPath;
-  // Allowed component extensions are jsx, tsx and json.
-  ['index.jsx', 'index.tsx', 'index.json'].some(item => {
-    const path = pathUtil.resolve(...pathSegments, item);
-    if (fs.existsSync(path)) {
-      componentPath = path;
-      return true;
-    }
-    return false;
-  });
-  return componentPath || null;
-};
-
-const cachedTemplates = [];
-
-const readTemplateFile = indexPath => {
-  if (cachedTemplates.includes(indexPath)) {
-    return cachedTemplates[indexPath];
-  }
-  if (!fs.existsSync(indexPath)) {
-    cachedTemplates[indexPath] = false;
-    return cachedTemplates[indexPath];
-  }
-  const contents = fs.readFileSync(indexPath);
-  try {
-    const parsedContent = JSON.parse(contents);
-    cachedTemplates[indexPath] = {
-      template: parsedContent['#template'],
-      subpage_template: parsedContent['#subpage_template'],
-      path: parsedContent.path,
-    };
-  } catch (exception) {
-    cachedTemplates[indexPath] = false;
-  }
-  return cachedTemplates[indexPath];
-};
-
-const findSubPageTemplateTemplate = (indexPath, basePath) => {
-  const templates = readTemplateFile(indexPath);
-  if (templates.subpage_template) return templates.subpage_template;
-  if (templates.template) return templates.template;
-  const parentPath = pathUtil.dirname(pathUtil.dirname(indexPath));
-  if (parentPath <= basePath) {
-    return '_default';
-  }
-  return findSubPageTemplateTemplate(`${parentPath}/index.json`, basePath);
-};
-
-const findTemplate = (indexPath, basePath, isFirst = true) => {
-  const templates = readTemplateFile(indexPath);
-  if (isFirst && templates.template) {
-    return templates.template;
-  }
-  if (!isFirst && templates.subpage_template) {
-    return templates.subpage_template;
-  }
-  const parentPath = pathUtil.dirname(pathUtil.dirname(indexPath));
-  if (parentPath <= basePath) {
-    return '_default';
-  }
-  return findTemplate(`${parentPath}/index.json`, basePath, false);
 };
 
 const createPagesFromFS = async ({ actions, graphql, getNode }) => {
