+++ packages/gatsby-theme-bodiless/package.json
@@ -76,7 +76,6 @@
     "debug": "^4.1.1",
     "dotenv": "^8.2.0",
     "fast-glob": "^3.2.5",
-    "find-up": "^4.1.0",
     "fs-extra": "^8.1.0",
     "gatsby-image": "^3.11.0",
     "gatsby-plugin-emotion": "^7.10.0",
@@ -93,10 +92,8 @@
     "gatsby-plugin-typescript": "^4.10.1",
     "gatsby-source-filesystem": "^4.11.1",
     "gatsby-transformer-sharp": "^4.10.0",
-    "git-branch": "^2.0.1",
     "glob": "^7.1.6",
     "informed": "^4.44.0",
-    "isomorphic-git": "^1.10.1",
     "lodash": "^4.17.19",
     "md5-file": "^5.0.0",
     "path-browserify": "^1.0.1",
