+++ sites/minimal-demo/package.json
@@ -21,6 +21,7 @@
     "@bodiless/fclasses": "^1.0.0-rc.37",
     "@bodiless/gatsby-theme-bodiless": "^1.0.0-rc.37",
     "@bodiless/layouts-ui": "^1.0.0-rc.37",
+    "@bodiless/page": "^1.0.0-rc.37",
     "@bodiless/richtext-ui": "^1.0.0-rc.37",
     "@bodiless/ui": "^1.0.0-rc.37",
     "autoprefixer": "^10.4.2",
