+++ sites/vital-demo/gatsby-config.js
@@ -4,7 +4,7 @@ const express = require('express');
 //   createDefaultContentPlugins,
 //   getSampleDefaultContentConfig,
 // } = require('@bodiless/gatsby-theme-bodiless/dist/DefaultContent');
-const { getDisabledPages } = require('@bodiless/components/lib/cjs/NodeApi');
+const { getDisabledPages } = require('@bodiless/page/lib/cjs/NodeApi');
 // const getSSIEntities = require('@bodiless/gatsby-plugin-ssi/read-ssi-entities');
 // const {
 //   getConfig: getSiteDefaultContentConfig,
